ECA-3685: Hopefully fixes this issue
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@19448 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ modules/cesecore-ejb/src/org/cesecore/certificates/ocsp/OcspResponseGeneratorSessionBean.java
@@ -1813,7 +1813,7 @@ public class OcspResponseGeneratorSessionBean implements OcspResponseGeneratorSe
         } else {
             dataMap.put(OcspKeyBinding.PROPERTY_RESPONDER_ID_TYPE, ResponderIdType.KEYHASH.name());
         }
-        dataMap.put(OcspKeyBinding.PROPERTY_MAX_AGE, (int)(OcspConfiguration.getMaxAge(CertificateProfileConstants.CERTPROFILE_NO_PROFILE)/1000L));
+        dataMap.put(OcspKeyBinding.PROPERTY_MAX_AGE, (long)(OcspConfiguration.getMaxAge(CertificateProfileConstants.CERTPROFILE_NO_PROFILE)/1000L));
         dataMap.put(OcspKeyBinding.PROPERTY_NON_EXISTING_GOOD, Boolean.valueOf(OcspConfiguration.getNonExistingIsGood()));
         dataMap.put(OcspKeyBinding.PROPERTY_NON_EXISTING_REVOKED, Boolean.valueOf(OcspConfiguration.getNonExistingIsRevoked()));
         dataMap.put(OcspKeyBinding.PROPERTY_UNTIL_NEXT_UPDATE, (long)(OcspConfiguration.getUntilNextUpdate(CertificateProfileConstants.CERTPROFILE_NO_PROFILE)/1000L));
